+++ tests/ApplicationTest.php
@@ -1,16 +1,16 @@
 <?php
 
-namespace ElectronTests;
+namespace TurbineTests;
 
 use League\Container\Container;
 use League\Event\Emitter;
 use League\Route\Http\Exception\NotFoundException;
 use League\Route\RouteCollection;
 use Monolog\Logger;
-use Electron;
-use Electron\Application;
-use ElectronTests\TestAsset\SharedTestController;
-use ElectronTests\TestAsset\TestController;
+use Turbine;
+use Turbine\Application;
+use TurbineTests\TestAsset\SharedTestController;
+use TurbineTests\TestAsset\TestController;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Zend\Diactoros\Response\TextResponse;