+++ CHANGELOG.md
@@ -1,12 +1,12 @@
-# Electron Changelog
+# Blast Turbine Changelog
 
 ## 1.0.0 (2016-03-04)
 
 ### Added
 
-- `Electron\Psr7\HttpKernelInterface` and `Electron\Psr7\TerminableInterface` port of symfony HttpKernelInterface for PSR-7 compatibility
+- `Turbine\Psr7\HttpKernelInterface` and `Turbine\Psr7\TerminableInterface` port of symfony HttpKernelInterface for PSR-7 compatibility
 - Add `zend/diactoros` for PSR-7 http support
-- provide compatibility with adapter `Electron\Symfony\HttpKernelAdapter` for StackPHP and other Symfony HttpKernelInterface implementations
+- provide compatibility with adapter `Turbine\Symfony\HttpKernelAdapter` for StackPHP and other Symfony HttpKernelInterface implementations
 
 ### Altered